7.9运行环境内容

+++ application/models/Friend_model.php
@@ -20,24 +20,75 @@ class Friend_model extends CI_Model{
 
     public function get_friend($openid,$page,$limit){
 
-        $data['result'] = $this->db//查询的字段
-        ->limit($limit,$page)
-        ->select(array('Friend.friendid','PersonalInfor.real_name','PersonalInfor.head_url','PersonalInfor.city','Work.company','Work.job'))
+        $friendid = $this->db//查询的字段
         ->where(array('Friend.openid'=>$openid,'Friend.state'=>'1'))//查询的条件
         ->from('Friend')//连表的主表
-        ->join('PersonalInfor','PersonalInfor.openid =  Friend.friendid')//连接表
-        ->join('Personal','Personal.openid =  Friend.friendid')//连接表
-        ->join('Work','Work.openid =  Friend.friendid')//连接表
-        ->get()->result_array();//语句查询，切记->result_array()；必须的
+        ->select('Friend.friendid')
+        ->limit($limit,$page)
+        ->get()
+        ->result_array();
 
-        $data['count'] = $this->db
-        ->select(array('Friend.friendid','PersonalInfor.real_name','PersonalInfor.head_url','PersonalInfor.city','Work.company','Work.job'))
-        ->where(array('Friend.openid'=>$openid,'Friend.state'=>'1'))
-        ->from('Friend')
-        ->join('PersonalInfor','PersonalInfor.openid =  Friend.friendid')
-        ->join('Personal','Personal.openid =  Friend.friendid')
-        ->join('Work','Work.openid =  Friend.friendid')
-        ->count_all_results();
+        $temp_count = $this->db//查询的字段
+        ->where(array('Friend.openid'=>$openid,'Friend.state'=>'1'))//查询的条件
+        ->from('Friend')//连表的主表
+        ->select('Friend.friendid')
+        ->get()
+        ->result_array();
+
+        $temp_data = array();
+
+        foreach ($friendid as $key){
+
+            $temp_openid = $key['friendid'];
+
+            $query_work = $this->db->get_where('Work',array('openid'=>$temp_openid))->result_array();
+
+            if($query_work===[]){
+                $temp = $this->db//查询的字段
+                ->select(array('PersonalInfor.openid','PersonalInfor.real_name','PersonalInfor.head_url','PersonalInfor.city','Education.school','Education.department'))
+                ->where(array('PersonalInfor.openid'=>$temp_openid))//查询的条件
+                ->from('PersonalInfor')//连表的主表
+                ->join('Education','Education.openid =  PersonalInfor.openid')//连接表
+                ->order_by('Education.end_year','DESC')
+                ->get()->result_array();//语句查询，切记->result_array()；必须的
+
+                if($temp[0] !== []){
+                    $temp[0]['company'] = $temp[0]['school'];
+                    $temp[0]['job'] = $temp[0]['department'];
+                    $temp[0]['friendid'] = $temp[0]['openid'];
+                }else{
+                    continue;
+                }
+
+                if($temp[0]['friendid']===null) continue;
+
+                array_push($temp_data,$temp[0]);
+
+            }else{
+                $temp = $this->db//查询的字段
+                ->select(array('PersonalInfor.openid','PersonalInfor.real_name','PersonalInfor.head_url','PersonalInfor.city','Work.company','Work.job'))
+                ->where(array('PersonalInfor.openid'=>$temp_openid))//查询的条件
+                ->from('PersonalInfor')//连表的主表
+                ->join('Work','Work.openid =  PersonalInfor.openid')//连接表
+                ->order_by('Work.end_year','DESC')
+                ->get()->result_array();//语句查询，切记->result_array()；必须的
+
+                if($temp[0] !== []){
+                    $temp[0]['friendid'] = $temp[0]['openid'];
+                }else{
+                    continue;
+                }
+
+                if($temp[0]['friendid']===null) continue;
+
+
+                array_push($temp_data,$temp[0]);
+            }
+
+        }
+
+        $data['result'] = $temp_data;
+        $data['count'] = count($temp_count,0);
 
         return $data;
     }
@@ -71,11 +122,32 @@ class Friend_model extends CI_Model{
 
     public function invite($openid,$friendid){
 
-        $this->db->delete('Friend',array('openid'=>$openid,'friendid'=>$friendid));
-        $this->db->insert('Friend',array('openid'=>$openid,'friendid'=>$friendid,'state'=>'3'));
+        $query = $this->db->get_where('PersonalInfor',array('openid'=>$friendid));
 
-        $this->db->delete('Friend',array('openid'=>$friendid,'friendid'=>$openid));
-        $this->db->insert('Friend',array('openid'=>$friendid,'friendid'=>$openid,'state'=>'2'));
+        if($query->result_array()===[]){
+            $data['status'] = '501';
+            $data['message'] = '用户不存在';
+        }else{
+            $query = $this->db
+                ->select('state')
+                ->get_where('Friend',array('openid'=>$openid,'friendid'=>$friendid))->result_array();
+
+            if($query === [] or $query[0]['state'] === '0'){
+                $this->db->delete('Friend',array('openid'=>$openid,'friendid'=>$friendid));
+                $this->db->insert('Friend',array('openid'=>$openid,'friendid'=>$friendid,'state'=>'3'));
+
+                $this->db->delete('Friend',array('openid'=>$friendid,'friendid'=>$openid));
+                $this->db->insert('Friend',array('openid'=>$friendid,'friendid'=>$openid,'state'=>'2'));
+
+                $data['status'] = '200';
+                $data['message'] = 'success';
+            }else{
+                $data['status'] = '502';
+                $data['message'] = '其他情况';
+            }
+        }
+
+        return $data;
 
     }
 
@@ -85,7 +157,6 @@ class Friend_model extends CI_Model{
         $this->db->insert('Friend',array('openid'=>$openid,'friendid'=>$friendid,'state'=>'1'));
 
         $this->db->delete('Friend',array('openid'=>$friendid,'friendid'=>$openid));
-        $this->db->insert('Friend',array('openid'=>$friendid,'friendid'=>$openid,'state'=>'1'));
         $this->db->insert('Friend',array('openid'=>$friendid,'friendid'=>$openid,'state'=>'5'));
 
     }
@@ -96,59 +167,9 @@ class Friend_model extends CI_Model{
         $this->db->insert('Friend',array('openid'=>$openid,'friendid'=>$friendid,'state'=>'0'));
 
         $this->db->delete('Friend',array('openid'=>$friendid,'friendid'=>$openid));
-        $this->db->insert('Friend',array('openid'=>$friendid,'friendid'=>$openid,'state'=>'0'));
         $this->db->insert('Friend',array('openid'=>$friendid,'friendid'=>$openid,'state'=>'4'));
 
     }
-
-
-    /*
-    *这四个暂时被getionform替代了
-    */
-    /*public function get_inviting($openid){
-
-        $data = $this->db
-        ->select(array('Friend.friendid','PersonalInfor.real_name','PersonalInfor.head_url','Friend.time','Friend.state'))
-        ->where(array('Friend.openid'=>$openid,'Friend.state'=>'3'))
-        ->from('Friend')
-        ->join('PersonalInfor','PersonalInfor.openid =  Friend.friendid')
-        ->get()->result_array();
-        return $data;
-    }
-
-    public function get_invited($openid){
-
-        $data = $this->db
-        ->select(array('Friend.friendid','PersonalInfor.real_name','PersonalInfor.head_url','Friend.time','Friend.state'))
-        ->where(array('Friend.openid'=>$openid,'Friend.state'=>'2'))
-        ->from('Friend')
-        ->join('PersonalInfor','PersonalInfor.openid =  Friend.friendid')
-        ->get()->result_array();
-        return $data;
-    }
-
-    public function get_refused($openid){
-
-        $data = $this->db
-        ->select(array('Friend.friendid','PersonalInfor.real_name','PersonalInfor.head_url','Friend.time','Friend.state'))
-        ->where(array('Friend.openid'=>$openid,'Friend.state'=>'4'))
-        ->from('Friend')
-        ->join('PersonalInfor','PersonalInfor.openid =  Friend.friendid')
-        ->get()->result_array();
-        return $data;
-    }
-
-    public function get_agreed($openid){
-
-        $data = $this->db
-        ->select(array('Friend.friendid','PersonalInfor.real_name','PersonalInfor.head_url','Friend.time','Friend.state'))
-        ->where(array('Friend.openid'=>$openid,'Friend.state'=>'5'))
-        ->from('Friend')
-        ->join('PersonalInfor','PersonalInfor.openid =  Friend.friendid')
-        ->get()->result_array();
-        return $data;
-    }*/
-
 
 
 }
